Refactor Options Dropdown to seperate components
