+++ app/components/Settings/OptionsDropdown/OptionsDropdown.react.js
@@ -1,16 +1,10 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
-import {flatten, keys} from 'ramda';
-import Select from 'react-select';
 import SQLOptions from './SQLOptions.react';
-import ESIndicesOptions from './ESDocsOptions.react';
+import ESIndicesOptions from './ESIndicesOptions.react';
+import ESDocsOptions from './ESDocsOptions.react';
+
 export default class OptionsDropdown extends Component {
-    constructor(props) {
-        super(props);
-        this.renderSQLOptions = this.renderSQLOptions.bind(this);
-        this.renderElasticsearchIndecies = this.renderElasticsearchIndecies.bind(this);
-        this.renderElasticsearchDocs = this.renderElasticsearchDocs.bind(this);
-    }
 
     static propTypes = {
         // See type of react-select's Select `value` property
@@ -22,10 +16,31 @@ export default class OptionsDropdown extends Component {
         setIndex: PropTypes.func
     }
 
+    /**
+     * Options Dropdown is an options drop down
+     * @param {object} props  - Component Properties
+     * @param {object} props.selectedTable - The selected table
+     * @param {object} props.tablesRequest - The Requested Table
+     * @param {function} props.setTable - The set table function
+     * @param {object} props.elasticsearchMappingsRequest - The ES Mapping Request
+     * @param {object} props.selectedTable - The selected table
+     * @param {object} props.selectedIndex - The Selected Index
+     * @param {function} props.setIndex - The Set the index
+     */
+    constructor(props) {
+        super(props);
+        this.renderSQLOptions = this.renderSQLOptions.bind(this);
+        this.renderElasticsearchIndecies = this.renderElasticsearchIndecies.bind(this);
+        this.renderElasticsearchDocs = this.renderElasticsearchDocs.bind(this);
+    }
+
     renderSQLOptions() {
         const {selectedTable, tablesRequest, setTable} = this.props;
 
-        return (<SQLOptions selectedTable={selectedTable} tablesRequest={tablesRequest} setTable={setTable} />);
+        return (<SQLOptions selectedTable={selectedTable}
+                            tablesRequest={tablesRequest}
+                            setTable={setTable}
+        />);
     }
 
     renderElasticsearchIndecies() {
@@ -35,41 +50,10 @@ export default class OptionsDropdown extends Component {
             selectedIndex
         } = this.props;
 
-        return (<ESIndicesOptions  elasticsearchMappingsRequest={EMR} 
-                                    setIndex={setIndex} 
-                                    selectedIndex={selectedIndex}/>);
-        /*if (!EMR.status) {
-            return null;
-        } else if (EMR.status === 'loading') {
-            return <div>{'Loading docs'}</div>;
-        } else if (EMR.status > 300) {
-            // TODO - Make this prettier.
-            return (
-                <div>
-                    <div>{'There was an error loading up your docs'}</div>
-                    <div style={{color: 'red'}}>{JSON.stringify(EMR)}</div>
-                </div>
-            );
-        } else if (EMR.status === 200) {
-            const indeciesList = keys(EMR.content);
-            if (indeciesList.length === 0) {
-                return <div>{'No docs found'}</div>;
-            }
-            return (
-                <div className={'dropdown'}
-                    id="test-table-dropdown"
-                >
-                    <Select
-                        options={indeciesList.map(t => ({label: t, value: t}))}
-                        value={selectedIndex}
-                        searchable={false}
-                        onChange={option => {
-                            setIndex(option.value);
-                        }}
-                    />
-                </div>
-            );
-        }*/
+        return (<ESIndicesOptions elasticsearchMappingsRequest={EMR}
+                                  setIndex={setIndex}
+                                  selectedIndex={selectedIndex}
+        />);
     }
 
     renderElasticsearchDocs() {
@@ -80,29 +64,11 @@ export default class OptionsDropdown extends Component {
             setTable
         } = this.props;
 
-        if (!selectedIndex) {
-            return null;
-        }
-
-        const tablesList = keys(EMR.content[selectedIndex].mappings);
-        if (tablesList.length === 0) {
-            return <div>{'No docs found'}</div>;
-        }
-
-        return (
-            <div className={'dropdown'}
-                id="test-table-dropdown"
-            >
-                <Select
-                    options={tablesList.map(t => ({label: t, value: t}))}
-                    value={selectedTable}
-                    searchable={false}
-                    onChange={option => {
-                        setTable(option.value);
-                    }}
-                />
-            </div>
-        );
+        return (<ESDocsOptions selectedTable={selectedTable}
+                                selectedIndex={selectedIndex}
+                                elasticsearchMappingsRequest={EMR}
+                                setTable={setTable}
+        />);
     }
 
     render() {
